pyproject.toml
[project]
name = "kbsb"
version = "3.4.3"
description = "API Website FRBE-KBSB-KSB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiocsv>=1.3.2",
    "aiofiles>=24.1.0",
    "async-lru>=2.0.5",
    "asyncclick>=8.1.8",
    "asyncssh>=2.21.0",
    "fastapi>=0.116.1",
    "google>=3.0.0",
    "google-api-python-client>=2.178.0",
    "google-auth>=2.40.3",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-storage>=3.2.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.4.0",
    "openpyxl>=3.1.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "reddevil",
    "requests>=2.32.4",
    "toml>=0.10.2",
    "unidecode>=1.4.0",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
reddevil = { path = "libs/reddevil-3.4.5-py3-none-any.whl" }


[dependency-groups]
dev = [
    "asyncio>=4.0.0",
    "black>=25.1.0",
    "freezegun>=1.5.5",
    "mypy>=1.17.1",
    "polyfactory>=2.22.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]
